+++ core/domain/interface/sale/sale_rep.go
@@ -11,5 +11,5 @@ package sale
 
 // 销售仓库
 type ISaleRep interface {
-	GetSale(mchId int) ISale
+	GetSale(mchId int64) ISale
 }
